requirements.txt
Flask==2.0.1
flask-cors
Werkzeug==2.2.3
numpy==1.24.4
pandas==1.5.3
pyrebase4==4.5.0
joblib==1.3.2
pickle-mixin==1.0.2
gunicorn==20.1.0
python-dotenv==1.0.0
firebase-admin
pyrebase4
requests
urllib3<2.0.0
scikit-learn